app.py
import os
import cv2
import time
import uuid
import sqlite3
import threading
from datetime import datetime
from collections import defaultdict

from fastapi import FastAPI, Request
from fastapi.responses import HTMLResponse, JSONResponse, Response
from fastapi.staticfiles import StaticFiles
from fastapi.templating import Jinja2Templates
from ultralytics import YOLO
from pydantic import BaseModel
from typing import Optional

# ── Serviços ────────────────────────────────────────────────────────────────
from services.event_repository import list_recent_events
from services.monitoring_agent import build_agent_messages, get_agent_status
from services import ollama_client
from services.config import AGENT_EVENT_LIMIT

# ── Configurações locais ─────────────────────────────────────────────────────
CAMERA_SOURCE = os.getenv("CAMERA_SOURCE", 0)
MODEL_PATH = "yolov8n.pt"
CONFIDENCE_THRESHOLD = 0.45
SAVE_DIR = "static/captures"
DB_PATH = "detections.db"

TARGET_CLASSES = {"person", "car", "motorcycle", "truck", "bus"}
MIN_CONSECUTIVE_FRAMES = 3
ALERT_COOLDOWN_SECONDS = 20

# ── App ───────────────────────────────────────────────────────────────────────
app = FastAPI(title="AgroVision AI")

os.makedirs("static", exist_ok=True)
os.makedirs("templates", exist_ok=True)
os.makedirs(SAVE_DIR, exist_ok=True)

app.mount("/static", StaticFiles(directory="static"), name="static")
templates = Jinja2Templates(directory="templates")

model = YOLO(MODEL_PATH)

last_frame = None
last_frame_lock = threading.Lock()
detection_state = defaultdict(int)
last_alert_time = defaultdict(lambda: 0.0)

# ── Schema do chat ────────────────────────────────────────────────────────────
class ChatRequest(BaseModel):
    message: str
    history: Optional[list] = []

# ── Banco de dados ────────────────────────────────────────────────────────────
def init_db():
    conn = sqlite3.connect(DB_PATH)
    cur = conn.cursor()
    cur.execute("""
        CREATE TABLE IF NOT EXISTS events (
            id TEXT PRIMARY KEY,
            event_time TEXT,
            label TEXT,
            confidence REAL,
            image_path TEXT
        )
    """)
    conn.commit()
    conn.close()


def save_event(event_id, label, confidence, image_path):
    conn = sqlite3.connect(DB_PATH)
    cur = conn.cursor()
    cur.execute("""
        INSERT INTO events (id, event_time, label, confidence, image_path)
        VALUES (?, ?, ?, ?, ?)
    """, (event_id, datetime.now().strftime("%Y-%m-%d %H:%M:%S"), label, confidence, image_path))
    conn.commit()
    conn.close()


def list_events(limit=50):
    conn = sqlite3.connect(DB_PATH)
    cur = conn.cursor()
    cur.execute("""
        SELECT id, event_time, label, confidence, image_path
        FROM events ORDER BY event_time DESC LIMIT ?
    """, (limit,))
    rows = cur.fetchall()
    conn.close()
    return [{"id": r[0], "event_time": r[1], "label": r[2], "confidence": r[3], "image_path": r[4]} for r in rows]

# ── Visão ─────────────────────────────────────────────────────────────────────
def draw_box(frame, x1, y1, x2, y2, label, conf):
    text = f"{label} {conf:.2f}"
    cv2.rectangle(frame, (x1, y1), (x2, y2), (0, 255, 0), 2)
    cv2.putText(frame, text, (x1, max(20, y1 - 10)), cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 255, 0), 2)


def should_alert(label):
    return (time.time() - last_alert_time[label]) > ALERT_COOLDOWN_SECONDS


def process_stream():
    global last_frame

    # Tenta converter para int se for webcam local (ex: "0")
    source = CAMERA_SOURCE
    try:
        source = int(source)
    except (ValueError, TypeError):
        pass

    cap = cv2.VideoCapture(source)
    if not cap.isOpened():
        print(f"[câmera] Erro ao abrir fonte: {source}")
        return

    print(f"[câmera] Iniciada com sucesso: {source}")

    while True:
        ok, frame = cap.read()
        if not ok:
            print("[câmera] Frame não lido, tentando reconectar...")
            time.sleep(int(os.getenv("CAMERA_RECONNECT_SECONDS", 5)))
            cap = cv2.VideoCapture(source)
            continue

        results = model(frame, conf=CONFIDENCE_THRESHOLD, verbose=False)
        found_labels_in_frame = set()
        best_conf_by_label = {}

        for result in results:
            boxes = result.boxes
            if boxes is None:
                continue
            for box in boxes:
                cls_id = int(box.cls[0].item())
                conf = float(box.conf[0].item())
                label = model.names[cls_id]
                if label not in TARGET_CLASSES:
                    continue
                found_labels_in_frame.add(label)
                if label not in best_conf_by_label or conf > best_conf_by_label[label]:
                    best_conf_by_label[label] = conf
                x1, y1, x2, y2 = map(int, box.xyxy[0].tolist())
                draw_box(frame, x1, y1, x2, y2, label, conf)

        for label in TARGET_CLASSES:
            if label in found_labels_in_frame:
                detection_state[label] += 1
            else:
                detection_state[label] = 0

        for label in found_labels_in_frame:
            if detection_state[label] >= MIN_CONSECUTIVE_FRAMES and should_alert(label):
                event_id = str(uuid.uuid4())[:8]
                filename = f"{datetime.now().strftime('%Y%m%d_%H%M%S')}_{label}_{event_id}.jpg"
                filepath = os.path.join(SAVE_DIR, filename)
                cv2.imwrite(filepath, frame)
                image_path = f"/static/captures/{filename}"
                confidence = best_conf_by_label.get(label, 0.0)
                save_event(event_id, label, confidence, image_path)
                last_alert_time[label] = time.time()
                print(f"[ALERTA] {label} detectado. Evidência salva em {filepath}")

        with last_frame_lock:
            last_frame = frame.copy()

        time.sleep(0.05)

# ── Startup ───────────────────────────────────────────────────────────────────
@app.on_event("startup")
def startup_event():
    init_db()
    thread = threading.Thread(target=process_stream, daemon=True)
    thread.start()

# ── Rotas ─────────────────────────────────────────────────────────────────────
@app.get("/", response_class=HTMLResponse)
def dashboard(request: Request):
    events = list_events(20)
    return templates.TemplateResponse("index.html", {"request": request, "events": events})


@app.get("/health")
def health():
    return {"status": "ok", "service": "AgroVision AI"}


@app.get("/events")
def get_events():
    return JSONResponse(content=list_events(50))


@app.get("/frame")
def get_frame():
    global last_frame
    with last_frame_lock:
        if last_frame is None:
            return JSONResponse(content={"message": "Ainda sem frame disponível."}, status_code=503)
        success, buffer = cv2.imencode(".jpg", last_frame)
        if not success:
            return JSONResponse(content={"message": "Erro ao converter frame."}, status_code=500)
        return Response(content=buffer.tobytes(), media_type="image/jpeg")


@app.post("/chat")
def chat(req: ChatRequest):
    """Recebe uma pergunta, monta o contexto com eventos recentes e consulta o agente."""
    events = list_recent_events(limit=AGENT_EVENT_LIMIT)
    messages = build_agent_messages(
        question=req.message,
        history=req.history or [],
        events=events,
    )
    answer = ollama_client.chat(messages)
    return {"answer": answer}


@app.get("/agent/status")
def agent_status():
    """Mostra o estado atual do agente: perfil, contexto e eventos carregados."""
    events = list_recent_events(limit=AGENT_EVENT_LIMIT)
    return get_agent_status(events)
